pyproject.toml
[build-system]
requires = ["setuptools>=65.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skillbench-pd"
version = "0.1.0"
description = "Benchmark for measuring progressive disclosure effects in Claude Skills"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "SkillBench-PD authors" }
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Topic :: Scientific/Engineering",
  "Topic :: Software Development :: Testing :: Benchmark"
]
dependencies = [
  "PyYAML>=6.0.1",
  "tabulate>=0.9.0",
  "matplotlib>=3.8.0"
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.26.0"]
dev = ["pytest>=7.4"]

[project.scripts]
skillbench-pd = "bench.cli:main"

[tool.setuptools.packages.find]
include = ["bench", "bench.*"]
